requirements.txt
astroid==2.3.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.2
PyYAML==5.3
requests==2.23.0
rope==0.14.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.8
wrapt==1.11.2
